Split SidebarMainSection into a parent class and two components
